# gin-blog-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin-vue-blog 后端接口文档。业务状态码 0 表示成功, 其他表示失败, HTTP 状态码统一为 200",
        "title": "gin-vue-blog API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/config": {
            "get": {
                "description": "获取博客配置, 优先读 Redis 缓存",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlogInfo"
                ],
                "summary": "获取博客配置",
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handle.Response-map_string_string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新博客配置, 同时清除 Redis 缓存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlogInfo"
                ],
                "summary": "更新博客配置",
                "parameters": [
                    {
                        "description": "博客配置",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handle.Response-any"
                        }
                    }
                }
            }
        },
...
        "/home": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "文章数, 用户数, 留言数, 访问量",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlogInfo"
                ],
                "summary": "获取后台首页信息",
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handle.Response-handle_BlogHomeVO"
                        }
                    }
                }
            }
        },
...
        "/setting/about": {
            "get": {
                "description": "获取关于我的内容",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlogInfo"
                ],
                "summary": "获取关于我",
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handle.Response-string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新关于我的内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlogInfo"
                ],
                "summary": "更新关于我",
                "parameters": [
                    {
                        "description": "关于我",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handle.AboutReq"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handle.Response-string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handle.AboutReq": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
...
        "handle.BlogHomeVO": {
            "type": "object",
            "properties": {
                "article_count": {
                    "description": "文章数量",
                    "type": "integer"
                },
                "message_count": {
                    "description": "留言数量",
                    "type": "integer"
                },
                "user_count": {
                    "description": "用户数量",
                    "type": "integer"
                },
                "view_count": {
                    "description": "访问量",
                    "type": "integer"
                },
                "view_trend": {
                    "description": "最近 viewTrendDays 天的访问量, 按日期升序, 没有数据的那天补 0",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handle.ViewTrendVO"
                    }
                },
                "visitor_area": {
                    "description": "访客地域分布, 按人数倒序, 最多 visitorAreaTop 项",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handle.VisitorAreaVO"
                    }
                }
            }
        },
...
        "handle.Response-any": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handle.Response-handle_BlogHomeVO": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handle.BlogHomeVO"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handle.Response-map_string_string": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/map_string_string"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handle.Response-string": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "type": "string"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handle.ViewTrendVO": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "当天访问量",
                    "type": "integer"
                },
                "date": {
                    "description": "2006-01-02",
                    "type": "string"
                }
            }
        },
        "handle.VisitorAreaVO": {
            "type": "object",
            "properties": {
                "area": {
                    "description": "省份, 取不到时是 \"未知\"",
                    "type": "string"
                },
                "count": {
                    "description": "独立访客数",
                    "type": "integer"
                }
            }
        },
...
        "map_string_string": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}